Updated libjpeg to version 9e (close #180)

+++ TextureLoader/src/JPEGCodec.c
@@ -209,7 +209,7 @@ ENCODE_JPEG_RESULT Diligent_EncodeJpeg(Uint8*     pSrcRGBPixels,
     /* Step 2: specify data destination (memory) */
     /* Note: steps 2 and 3 can be done in either order. */
     unsigned char* mem      = NULL;
-    unsigned long  mem_size = 0;
+    size_t         mem_size = 0;
     jpeg_mem_dest(&cinfo, &mem, &mem_size);
 
     /* Step 3: set parameters for compression */
